// node/typings/promotions.ts
export interface GetAllBenefitsResponseScope {
  allCatalog: boolean
  skus: number
  skusAreInclusive: boolean
  products: number
  productsAreInclusive: boolean
  collections: number
  collectionsAreInclusive: boolean
  categories: number
  categoriesAreInclusive: boolean
  brands: number
  brandsAreInclusive: boolean
  sellers: number
  sellersAreInclusive: boolean
}

export interface GetAllBenefitsResponseItem {
  idCalculatorConfiguration: string
  lastModifiedUtc: string
  name: string
  beginDate: string
  endDate: string
  isActive: boolean
  description: string
  type: string
  utmSource: string
  utmCampain: string
  utmiCampaign: string
  status: string
  percentualTax: number
  isArchived: boolean
  hasMaxPricePerItem: boolean
  isTax: boolean
  Campaigns: string[]
  conditionsIds: string[]
  activateGiftsMultiplier: boolean
  effectType: string
  scope: GetAllBenefitsResponseScope
  maxUsage: number
  idsSalesChannel: string[]
  areSalesChannelIdsExclusive: boolean
}